refactor(api): move tests
